use SDOs to send commands
The Roboteq CANOpen implementation seem broken, and I can't manage
to get both TPDO and RPDO working at the same time. Use SDOs to
send commands instead of RPDOs, and keep TPDOs for status report.

To make this work, we must FIRST read a TPDO before we start
using the SDOs to send commands. Without this sequence, the
SDO writing will "gobble" the PDOs and no status is ever sent.