output the controller status flags
Note that this setup is *very* specific to our (Tidewise's) use
case. The driver uses one PDO per channel (we use only one) plus
two for the status, and we only need one analog input (half
of a PDO).

Anything more demanding won't configure as it would need more PDOs.
The configuration interface would need to be updated to allow
selecting which analogs go in PDOs, and the controller status should
be using SDOs for low-priority values (e.g. voltages and temperatures)